graphql_sub/graphql-social-network/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asgiref==3.10.0
asttokens==2.4.1
blinker==1.9.0
click==8.3.1
comm==0.2.2
contourpy==1.3.3
cross-web==0.4.1
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
Django==4.2.25
django-admin-interface==0.30.1
django-colorfield==0.14.0
executing==2.1.0
fastapi==0.128.0
Flask==3.1.2
fonttools==4.61.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.1.7
neo4j==6.1.0
nest-asyncio==1.6.0
numpy==2.3.5
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.18.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.22
python-slugify==8.0.4
pytz==2026.1.post1
pyzmq==26.2.0
Rx==3.2.0
six==1.17.0
SQLAlchemy==2.0.46
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.50.0
strawberry-graphql==0.308.2
text-unidecode==1.3
tornado==6.4.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
wcwidth==0.2.13
Werkzeug==3.1.5
whitenoise==6.11.0
